1.20.1 - it compiles and create a world (with a lot of errors)

+++ src/main/java/de/melanx/skyblockbuilder/data/Team.java
@@ -430,14 +430,14 @@ public class Team {
         this.possibleSpawns.clear();
         for (Tag pos : spawns) {
             CompoundTag posTag = (CompoundTag) pos;
-            this.possibleSpawns.add(new BlockPos(posTag.getDouble("posX"), posTag.getDouble("posY"), posTag.getDouble("posZ")));
+            this.possibleSpawns.add(new BlockPos(posTag.getInt("posX"), posTag.getInt("posY"), posTag.getInt("posZ"))); // todo double
         }
 
         ListTag defaultSpawns = nbt.getList("DefaultSpawns", Tag.TAG_COMPOUND);
         this.defaultPossibleSpawns.clear();
         for (Tag pos : defaultSpawns) {
             CompoundTag posTag = (CompoundTag) pos;
-            this.defaultPossibleSpawns.add(new BlockPos(posTag.getDouble("posX"), posTag.getDouble("posY"), posTag.getDouble("posZ")));
+            this.defaultPossibleSpawns.add(new BlockPos(posTag.getInt("posX"), posTag.getInt("posY"), posTag.getInt("posZ"))); // todo double
         }
 
         ListTag joinRequests = nbt.getList("JoinRequests", Tag.TAG_COMPOUND);
